+++ src/components/cards/contentCard/ContentCard.style.ts
