+++ src/components/headers/closeButtonHeader/CloseButtonHeader.style.ts
