+++ src/components/lists/categoryList/CategoryList.style.ts
