+++ src/screens/contentList/ContentList.tsx
@@ -6,10 +6,10 @@ import { ListRenderItem } from 'react-native'
 
 import { getContentList } from '@/apis/fakeServerApis'
 import { defaultImages } from '@/assets'
-import ContentCard from '@/components/ContentCard/ContentCard'
+import ContentCard from '@/components/cards/contentCard/ContentCard'
 import DefaultDialog from '@/components/dialogs/defaultDialog/DefaultDialog'
 import TwoButtonDialog from '@/components/dialogs/twoButtonDialog/TwoButtonDialog'
-import DefaultHeader from '@/components/header/defaultHeader/DefaultHeader'
+import DefaultHeader from '@/components/headers/defaultHeader/DefaultHeader'
 import Popup from '@/components/popup/Popup'
 import i18n from '@/locales'
 import { PopupMenu } from '@/models/PopupMenu'
