+++ src/screens/report/Report.tsx
@@ -7,7 +7,7 @@ import BottomSheet from '@/components/bottomSheet/BottomSheet'
 import { BoxButton } from '@/components/buttons/boxButton/BoxButton'
 import RadioButton from '@/components/buttons/radioButton/RadioButton'
 import DefaultDialog from '@/components/dialogs/defaultDialog/DefaultDialog'
-import { LeftButtonHeader } from '@/components/header/leftButtonHeader/LeftButtonHeader'
+import { LeftButtonHeader } from '@/components/headers/leftButtonHeader/LeftButtonHeader'
 import reportMenuConfig from '@/configs/reportMenuConfig.json'
 import i18n from '@/locales'
 import { ReportMenu } from '@/models/ReportMenu'
